move BinaryGenerator to Binary folder
